books being overwritten this.setstate
issue with search.js where dropdown does not work:
currently search.js had its own state that was being updated, but the state in books which was used for the comparison in app.js changeShelfes, was not being updated.
Current issue: all data is being removed as I am setting this.setState to book, but overwriting books.